fix snapshots
